Work so far
HF is the header and footer. The header still needs to be worked on (check beginning comments in hf.html file), footer is yet to become WIP.

The styles.css file has all the necessary css for the header so far. Please check it out and add to it - COMMENT WHICH PART IS YOURS TO AVOID CONFUSION OR CLUSTERFUCKERY

Orange.jpg is a test file for the background. Basically irrelevant :-)