give menu buttons some function
Currently they just jump to the specific part without any scrolling animation (I don't know enough jQuery to do that right now).
This commit also gives the footer an id, which was the easiest way I could think of to get the Contact button to scroll all the way down.